requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pandas==2.2.3
social-auth-core
pyotp
pydantic-settings==2.9.1
